Netflow data collection support
This patch adds support for getting netflow data from containers.
The fprobe tool is started and periodically sends netflow to a
local 'socket-datacollector' that writes the data into files.

The fprobe plugin also picks up the files written by the
'socket-datacollector' and erases them.

We run a cleanup every 30 seconds where we check the currently
available interfaces versus those for which we have started fprobes.
If interfaces on the host are found to be missing (container ended)
we remove all those files belonging to the container that had this
interface.

This patch adds the following dependencies:
- fprobe (apt-get -y install fprobe)
- socket-datacollector (available from Stefan's internal github page)

Signed-off-by: Stefan Berger <stefanb@linux.vnet.ibm.com>