Rework how we check for interfaces for which fprobes are running
Rather than keeping a list of interfaces for which fprobes are running,
use a map that holds the list of interfaces as keys and the fprobe's
process ID as the value. This replaces the current set().

Every time we crawl  a container and determine which interfaces need
to have fprobes running on, check whether there is an entry in the map
or the process ID in the key is still running or representing a fprobe
(which could have changed due to pid-reuse). After starting the fprobe
successfully, record the process ID of the PID.

When an fprobes disappeared on an interface it will be restarted upon
the next crawl().

Signed-off-by: Stefan Berger <stefanb@linux.vnet.ibm.com>