Allow to specify the maximum number of file descriptors to close
To avoid leaking of file descriptors to started child processes
we close all the ones we don't want to pass. This works efficiently
for as long as the max. number of file descriptors the process
can have is low, such as 1024. When running the crawler on a Ubuntu
host, this number is 1024 (/proc/<pid>/limits). When running it inside
a container, the number may for example be 1048576. So, we introduce
a parameter to the start_child function that allows us to artificially
limit that number for a process, like the crawler, that doesn't have so
many file descriptors open. We pass 128, which is sufficiently high.

Signed-off-by: Stefan Berger <stefanb@linux.vnet.ibm.com>