Switch from fprobe to softflowd
Softflowd seems to be the more reliable tool compared to fprobe.
Use softflowd now as a more or less straight replacement. Change
some of the comments in the code to not reflect the tool fprobe
anymore but either refer to a 'flow probe' or 'softflowd' if it's
about the program name.

Only use one timeout parameter for softflowd, the 'maxlife' timeout
parameter that expires all flows after a few seconds. This avoids
having to provide at least 6 individual parmeters for controlling
timeouts for TCP, UDP, etc. Since softflowd produces netflow V{1,5,9},
we need to adjust this as well from fprobe's V{1,5,7}.

Address some pylint complaints about import order.

Make the building of the container more efficient by reversing the
order of commands.

Signed-off-by: Stefan Berger <stefanb@linux.vnet.ibm.com>