fprobe: Delay getting interface IP address
In rare cases we seem to enter a container and find an interface
before Docker has assigned the IP address to it. To avoid
starting the socket-datacollector without IP address information
we delay getting the IP address inside the container, in case
none was found ,and try again after a 1ms delay. We repeat up to
4 times.

In case we still end up without an IP address, we don't start
the flow probe but try again at the next crawl.

Signed-off-by: Stefan Berger <stefanb@linux.vnet.ibm.com>