Fix the Travis build by requiring a specific version of pytest
A more recent version of pytest is being installed by not being
specific about its version when installing pytest-cov, which will
put pytest as a dependency. The later install of requirements-test.txt
does not solve that problem, either. So we install the pytest
with the specific version 2.9.2 when installing pytest-cov with specific
version 2.3.0.

Signed-off-by: Stefan Berger <stefanb@linux.vnet.ibm.com>