ctprobe: Do not use socket-datacollector anymore
Do not use the socket-datacollector anymore but use the
conntrackprobe's file+json:// URL type to have it write the
same type of files. Conntrackprobe also resolves {timestamp}
and {ifname} to their respective values so that filenames
allow to distinguish the sources (ifname) and multiple
sequential files with data are written (timestamp).

Using conntrackprobe for writing the files makes the plugin
simpler.

Signed-off-by: Stefan Berger <stefanb@linux.vnet.ibm.com>